+++ client/pages/page-1/module-1/components/com-2/index.js
@@ -1,13 +1,9 @@
 import React, {Component} from 'react';
 import PropTypes from 'prop-types';
 import {openToast} from '../../../../../common/action/toast';
-import {helloModule1} from '../../action';
+import callApi from '../../../../../utils/fetch';
 
 export class Com2 extends Component {
-  static propTypes = {
-    module1: PropTypes.object,
-  };
-
   static contextTypes = {
     dispatch: PropTypes.func,
   };
@@ -16,8 +12,12 @@ export class Com2 extends Component {
     this.context.dispatch(openToast('Module-1 Com-2'));
   };
 
-  handleHello = () => {
-    this.context.dispatch(helloModule1('您好，欢迎您来到这里'));
+  handleFetch = (url) => {
+    return () => {
+      callApi({
+        url
+      });
+    };
   };
 
   render() {
@@ -26,9 +26,12 @@ export class Com2 extends Component {
         <h3 className="m-y-4">Module-1 Com-2</h3>
         <button className="btn btn-primary btn-raised" onClick={this.toast}>Toast</button>
 
-        <div className="m-t-4">
-          <button className="btn btn-secondary btn-raised" onClick={this.handleHello}>Hello World</button>
-          <p className="theme-secondary m-t-2">{this.props.module1.get('content')}</p>
+        <div className="m-t-5">
+          <p>多接口数据请求，返回数据请在控制台中查看</p>
+          <div className="btn-group m-t-4">
+            <button className="btn btn-primary btn-raised" onClick={this.handleFetch('page-1/sync-list')}>同步</button>
+            <button className="btn btn-secondary btn-raised" onClick={this.handleFetch('page-1/async-list')}>异步</button>
+          </div>
         </div>
       </div>
     );