+++ server/api-url/page-1.js
@@ -2,7 +2,8 @@
 const {serverUrl} = require('../config');
 
 const apiUrl = {
-  list: `${serverUrl}list`,
+  list1: `${serverUrl}list1`,
+  list2: `${serverUrl}list2`,
 };
 
 module.exports = apiUrl;